review: strip godoc comments and fold tests into authn_test.go
Addresses review feedback on #113.

- Move the three verifier tests from internal/client/verifier_test.go
  into internal/client/authn_test.go (no test logic changed; docstrings
  dropped to match the surrounding file's style) and delete the now-empty
  verifier_test.go.
- Remove the godoc comments on:
    - pkg/goidc/model.go: VerifyClientSecretFunc
    - internal/oidc/config.go: VerifyClientSecretFunc field
    - internal/oidc/context.go: Context.VerifyClientSecret method
    - pkg/provider/provider.go: defaultVerifyClientSecretFunc
  per maintainer preference for a terser surface.
